+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.3.0] -  2025-03-31
+
+- Add EBS volume backup.
+- Add EBS volume restore.
+
 ## [2.2.0] - 2025-03-11
 
 - The open vpn version was updated to 2.6.12 and the Alpine version of the container was updated to 3.21.3.